+++ Questless/Questless/src/quantities/wall_time.hpp
@@ -10,27 +10,27 @@
 
 namespace ql {
 	//! Wall time duration in seconds.
-	using sec = cancel::quantity<double, cancel::unit_t<struct sec_tag>>;
+	using sec = cancel::quantity<float, cancel::unit_t<struct sec_tag>>;
 	constexpr sec operator"" _s(long double value) {
-		return sec{static_cast<double>(value)};
+		return sec{static_cast<float>(value)};
 	}
 
 	//! Wall time duration in milliseconds.
-	using msec = cancel::quantity<double, cancel::unit_t<struct msec_tag>>;
+	using msec = cancel::quantity<float, cancel::unit_t<struct msec_tag>>;
 	constexpr msec operator"" _ms(long double value) {
-		return msec{static_cast<double>(value)};
+		return msec{static_cast<float>(value)};
 	}
 
 	//! Wall time rate in inverse seconds.
 	using per_sec = cancel::inverse_t<sec>;
 	constexpr per_sec operator"" _hz(long double value) {
-		return per_sec{static_cast<double>(value)};
+		return per_sec{static_cast<float>(value)};
 	}
 
 	//! Wall time rate in inverse milliseconds.
 	using per_msec = cancel::inverse_t<msec>;
 	constexpr per_msec operator"" _khz(long double value) {
-		return per_msec{static_cast<double>(value)};
+		return per_msec{static_cast<float>(value)};
 	}
 
 	constexpr auto ms_per_sec = 1000.0_ms / 1.0_s;
@@ -40,13 +40,13 @@ namespace ql {
 	using clock = std::chrono::steady_clock;
 
 	//! Wall time in std::chrono seconds.
-	using chrono_sec = std::chrono::duration<double>;
+	using chrono_sec = std::chrono::duration<float>;
 
 	//! The game's target frames per second.
 	static constexpr auto target_frame_rate = 60.0_hz;
 
 	//! The target duration of a single frame; inverse of FPS.
-	static constexpr auto target_frame_duration = 1.0 / target_frame_rate;
+	static constexpr auto target_frame_duration = 1.0f / target_frame_rate;
 
 	//! Converts from a seconds quantity to std::chrono seconds. Useful for interacting with the Standard Library.
 	constexpr auto to_chrono_sec(sec sec) {
